+++ Editor/Scripts/Drawers/LayerAttributeDrawer.cs
@@ -1,0 +1,42 @@
+/*
+ * File Name: LayerAttributeDrawer.cs
+ * Description: This script is for ...
+ * 
+ * Author(s): Kokowolo, Will Lacey
+ * Date Created: April 8, 2024
+ * 
+ * Additional Comments:
+ *      File Line Length: 140
+ */
+
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+using UnityEditor;
+
+namespace Kokowolo.Utilities.Editor
+{
+    [CustomPropertyDrawer(typeof(LayerAttribute))]
+    public class LayerAttributeDrawer : PropertyDrawer
+    {
+        /************************************************************/
+        #region Fields
+
+        #endregion
+        /************************************************************/
+        #region Properties
+
+        #endregion
+        /************************************************************/
+        #region Functions
+
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            property.intValue = EditorGUI.LayerField(position, label, property.intValue);
+        }
+
+        #endregion
+        /************************************************************/
+    }
+}