+++ Runtime/Scripts/Attributes/LayerAttribute.cs
@@ -1,0 +1,22 @@
+/*
+ * File Name: LayerAttribute.cs
+ * Description: This script is for ...
+ * 
+ * Author(s): Kokowolo, Will Lacey
+ * Date Created: April 8, 2024
+ * 
+ * Additional Comments:
+ *      File Line Length: 140
+ */
+
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Kokowolo.Utilities
+{
+    public class LayerAttribute : PropertyAttribute
+    {
+
+    }
+}